fix: missing gql api change in mutation
